rm unchecked warnings in tools
