+++ api/tools/src/main/java/org/pmiops/workbench/tools/factories/ToolsRawlsApiClientFactory.java
@@ -5,20 +5,13 @@ import org.pmiops.workbench.rawls.ApiClient;
 import org.pmiops.workbench.rawls.api.BillingV2Api;
 import org.pmiops.workbench.rawls.api.WorkspacesApi;
 
-public abstract class ToolsRawlsApiClientFactory {
+public abstract class ToolsRawlsApiClientFactory extends ToolsApiClientFactory {
 
   protected ApiClient apiClient;
 
   protected ToolsRawlsApiClientFactory(ApiClient apiClient) {
     this.apiClient = apiClient;
   }
-
-  protected static final String[] FC_SCOPES =
-      new String[] {
-        "https://www.googleapis.com/auth/userinfo.profile",
-        "https://www.googleapis.com/auth/userinfo.email",
-        "https://www.googleapis.com/auth/cloud-billing"
-      };
 
   public WorkspacesApi workspacesApi() throws IOException {
     WorkspacesApi api = new WorkspacesApi();
